Fix review comments
- Use label to find items between upgrade
- Move blazegraph function to default-functions.js
- Move getElementByURI to blazegraph.js helper
- Split the QS post upgrade to run pre and post
