+++ Docker/test/selenium/wdio.conf.js
@@ -117,12 +117,14 @@ exports.config = {
 
 		pre_upgrade: [
 			'./specs/repo/api.js',
-			'./specs/upgrade/pre-upgrade.js'
+			'./specs/upgrade/pre-upgrade.js',
+			'./specs/upgrade/queryservice-pre-and-post-upgrade.js'
 		],
 
 		post_upgrade: [
 			'./specs/repo/api.js',
 			'./specs/upgrade/post-upgrade.js',
+			'./specs/upgrade/queryservice-pre-and-post-upgrade.js',
 			'./specs/upgrade/queryservice-post-upgrade.js'
 		]
